+++ Sources/MapboxNavigation/Style.swift
@@ -44,6 +44,8 @@ open class Style: NSObject {
      */
     @objc open var previewMapStyleURL = URL(string: StyleURI.navigationDay.rawValue)!
     
+    var traitCollection: UITraitCollection = UITraitCollection(userInterfaceIdiom: .phone)
+    
     /**
      Applies the style for all changed properties.
      */