+++ Tests/MapboxNavigationTests/CarPlayCompassViewSnapshotTests.swift
@@ -4,7 +4,15 @@ import SnapshotTesting
 import TestHelper
 @testable import MapboxNavigation
 
+class CarPlayCompassViewMock: CarPlayCompassView {
+    
+    override var traitCollection: UITraitCollection {
+        return UITraitCollection(userInterfaceIdiom: .carPlay)
+    }
+}
+
 class CarPlayCompassViewSnapshotTests: TestCase {
+    
     private let styles = [DayStyle(), NightStyle()]
     
     override func setUp() {
@@ -14,21 +22,26 @@ class CarPlayCompassViewSnapshotTests: TestCase {
     
     func testCarPlayCompassView() {
         for style in styles {
+            // `StyleManager` switches between user interface styles depending on platform, on
+            // which it was created (to prevent global appearance updates of UI components that are
+            // used on both phone and CarPlay).
+            // In this case trait collection is changed directly.
+            style.traitCollection = UITraitCollection(userInterfaceIdiom: .carPlay)
+            
             let stackView = UIStackView(orientation: .vertical, spacing: 5, autoLayout: true)
             style.apply()
             
             let horizontalStackView = UIStackView(orientation: .horizontal, spacing: 2, autoLayout: true)
             
             for course in stride(from: 0, to: 360, by: 45) {
-                let compassView = CarPlayCompassView(frame: .zero)
-                compassView.isHidden = false
-                compassView.course = CLLocationDirection(course)
-                horizontalStackView.addArrangedSubview(compassView)
+                let carPlayCompassViewMock = CarPlayCompassViewMock(frame: .zero)
+                carPlayCompassViewMock.isHidden = false
+                carPlayCompassViewMock.course = CLLocationDirection(course)
+                horizontalStackView.addArrangedSubview(carPlayCompassViewMock)
             }
             
             stackView.addArrangedSubview(horizontalStackView)
             assertImageSnapshot(matching: stackView, as: .image(precision: 0.95))
         }
-
     }
 }